% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_xl.R
\name{import_cl_excel}
\alias{import_cl_excel}
\title{Import de tous les fichiers Excel d'un dossier et agrégation dans un seul tableau}
\usage{
import_cl_excel(path = getwd(), ext = "xlsx", donnees_manquantes = FALSE)
}
\arguments{
\item{path}{Emplacement du dossier}

\item{ext}{Extension des fichiers (xls ou xlsx)}
}
\description{
Import de tous les fichiers Excel d'un dossier et agrégation dans un seul tableau
}
