% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pdf.R
\name{import_cl_pdf}
\alias{import_cl_pdf}
\title{Import de tous les fichiers Excel d'un dossier et agrégation dans un seul tableau}
\usage{
import_cl_pdf(path = getwd(), donnees_manquantes = FALSE)
}
\arguments{
\item{path}{Emplacement du dossier}
}
\description{
Import de tous les fichiers Excel d'un dossier et agrégation dans un seul tableau
}
