% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indic_mammites_mois.R
\name{indic_mam_mois}
\alias{indic_mam_mois}
\title{Calcul des indicateurs mammites mensuels entre 2 dates}
\usage{
indic_mam_mois(ctrl_indiv = data.frame(), date_debut = NULL, date_fin = NULL)
}
\arguments{
\item{ctrl_indiv}{Tableau de données contrôle laitier avec une ligne par vache-date de contrôle}

\item{date_debut}{Date de début de la période à prendre en compte (Format aaaa-mm-jj)}

\item{date_fin}{Date de fin de la période à prendre en compte (Format aaaa-mm-jj)}
}
\description{
Calcul des indicateurs mammites mensuels entre 2 dates
}
